Add files via upload
This code is extremely similar to how the time series for Saildrone was plotted. These plots correspond to the 2017/2018 1005/1006 Saildrone Mission. The TAO data downloaded corresponds to the dates Saildrone was deployed, which is why indexing the data was not needed. Will most likely be uploading every python file, but if not, the TAO data used needs to be indexed to the specific Saildrone mission being targeted. I did this to make aligning data easier and make it more organized in my brain :)

This code looks at the wind components for 2017-2018 SD1005/1006 at 0,110W; 0,140W; 0,170W, 0,165E; and 0,125W. This code was then replicated for the remaining Saildrone Missions (+4, 5 mission altogether).